fix(arenabench): split CloudExecutor into cloud_executor.py to clear the ratchet
The #3255 refusal check added ~40 lines to arenabench/arenabench/cloud.py,
crossing it over the 1500-line file-size ratchet for the first time (1540
lines, never grandfathered) and failing both the "file size ratchet" and
"fmt + clippy + test" checks on PR #3300.

CloudExecutor (the boto3 adapter, ~480 lines) is the largest cohesive chunk
in the file and the one piece that is not a pure decision function, so it
moves to a sibling module, arenabench/arenabench/cloud_executor.py, matching
the existing cloud_merge.py/cloud_watch.py split pattern. cloud.py
re-exports it (`from .cloud_executor import CloudExecutor`) so every existing
import site (`arenabench.cloud.CloudExecutor` in tests, cli.py, gate_watch.py
docstrings) keeps working unchanged. The import is placed after the pure
functions/constants CloudExecutor depends on, which is what lets the
two-way module relationship (cloud_executor imports decisions from cloud;
cloud imports the class back from cloud_executor) resolve without a
circular-import error.

cloud.py is now 1064 lines; cloud_executor.py is 544. Also drops two imports
from cloud.py that became unused once the class moved (`json`, `is_safe_ref`).
